requirements.txt
asn1crypto==1.2.0
astroid==2.3.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorz==1.0.3
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
haishoku==1.1.8
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
Pillow==6.2.1
pycosat==0.6.3
pycparser==2.19
pylint==2.4.4
pyOpenSSL==19.1.0
PySocks==1.7.1
pywal==3.3.0
requests==2.22.0
ruamel-yaml==0.15.46
scipy==1.3.3
six==1.13.0
tqdm==4.40.0
urllib3==1.25.7
wrapt==1.11.2
yapf==0.28.0
